codecheck

+++ service/hnp/base/hnp_file.c
@@ -178,34 +178,20 @@ int GetRealPath(char *srcPath, char *realPath)
     return 0;
 }
 
-int HnpDeleteFolder(const char *path)
+static int HnpDeleteAllFileInPath(const char *path, DIR *dir)
 {
-    int ret = 0;
-#ifdef _WIN32
-    return ret;
-#else
     struct dirent *entry;
     struct stat statbuf;
     char filePath[MAX_FILE_PATH_LEN];
-
-    if (access(path, F_OK) != 0) {
-        return 0;
-    }
-    DIR *dir = opendir(path);
-    if (dir == NULL) {
-        HNP_LOGE("delete folder open dir=%s unsuccess ", path);
-        return HNP_ERRNO_BASE_DIR_OPEN_FAILED;
-    }
 
     while ((entry = readdir(dir)) != NULL) {
         if ((strcmp(entry->d_name, ".") == 0) || (strcmp(entry->d_name, "..") == 0)) {
             continue;
         }
 
-        ret = sprintf_s(filePath, MAX_FILE_PATH_LEN, "%s/%s", path, entry->d_name);
+        int ret = sprintf_s(filePath, MAX_FILE_PATH_LEN, "%s/%s", path, entry->d_name);
         if (ret < 0) {
             HNP_LOGE("delete folder sprintf path[%s], file[%s] unsuccess.", path, entry->d_name);
-            closedir(dir);
             return HNP_ERRNO_BASE_SPRINTF_FAILED;
         }
 
@@ -217,20 +203,42 @@ int HnpDeleteFolder(const char *path)
             /* 如果是文件夹，递归删除 */
             ret = HnpDeleteFolder(filePath);
             if (ret != 0) {
-                closedir(dir);
                 return ret;
             }
         } else {
             ret = unlink(filePath);
             if (ret != 0) {
-                closedir(dir);
                 HNP_LOGE("delete file unsuccess.ret=%d, path=%s, errno=%d", ret, filePath, errno);
                 return HNP_ERRNO_BASE_UNLINK_FAILED;
             }
         }
     }
 
+    return 0;
+}
+
+int HnpDeleteFolder(const char *path)
+{
+    int ret = 0;
+#ifdef _WIN32
+    return ret;
+#else
+
+    if (access(path, F_OK) != 0) {
+        return 0;
+    }
+    DIR *dir = opendir(path);
+    if (dir == NULL) {
+        HNP_LOGE("delete folder open dir=%s unsuccess ", path);
+        return HNP_ERRNO_BASE_DIR_OPEN_FAILED;
+    }
+
+    ret = HnpDeleteAllFileInPath(path, dir);
     closedir(dir);
+    if (ret != 0) {
+        return ret;
+    }
+
     ret = rmdir(path);
     if (ret != 0) {
         HNP_LOGE("rmdir path unsuccess.ret=%d, path=%s, errno=%d", ret, path, errno);
